+++ migrations/frontend/1528395909_undo_apidocs_oob_migration.down.sql
@@ -1,0 +1,5 @@
+BEGIN;
+
+-- Nothing to do, the up migration undid changes previously made.
+
+COMMIT;